Revert "Revert "staging: android: binder: move to the "real" part of the kernel""
This reverts commit 1d07e6dd8f9100a81da29351a7211246dc64d95a.
