+++ include/uapi/linux/android/Kbuild
@@ -1,0 +1,2 @@
+# UAPI Header export list
+header-y += binder.h
